Update: Tingle.js を使用して実装した Permission 確認モーダル (iOS) を別ファイルに移動
