+++ src/components/organisms/IOSPermissionsCheckModal/index.ts
@@ -1,0 +1,1 @@
+export { IOSPermissionsCheckModal } from "./IOSPermissionsCheckModal";
