+++ src/components/organisms/MobileSensorData/MobileSensorData.tsx
@@ -1,8 +1,10 @@
 import React, { useEffect, useState, useMemo } from "react";
-import tingle from "tingle.js";
+
+// Components
 import { CardArticle } from "../../molecules/CardArticle";
 import { Section } from "../../molecules/Section";
 import { List } from "../../molecules/Lists";
+import { IOSPermissionsCheckModal } from "../IOSPermissionsCheckModal";
 
 type Orientation = DeviceOrientationEvent | null;
 type Acceleration = DeviceMotionEventAcceleration | null;
@@ -15,71 +17,6 @@ export const MobileSensorData: React.VFC = () => {
   const [accelerationIncludingGravity, setAccelerationIncludingGravity] =
     useState<AccelerationIncludingGravity>(null);
   const [rotationRate, setRotationRate] = useState<RotationRate>(null);
-
-  useEffect(() => {
-    // FIXME: iOS対応の為必要とする
-    // https://qiita.com/Toya-Onodera/items/fe06000a7909797fb5b4
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    if (
-      typeof (DeviceMotionEvent as any).requestPermission === "function" &&
-      sessionStorage.getItem("isPermission") === null
-    ) {
-      const tingleLinkElement = document.createElement("link");
-      tingleLinkElement.rel = "stylesheet";
-      tingleLinkElement.href =
-        "https://cdnjs.cloudflare.com/ajax/libs/tingle/0.15.3/tingle.min.css";
-      document.head.appendChild(tingleLinkElement);
-
-      tingleLinkElement.onload = () => {
-        const tingleScriptElement = document.createElement("script");
-        tingleScriptElement.src =
-          "https://cdnjs.cloudflare.com/ajax/libs/tingle/0.15.3/tingle.min.js";
-        document.body.appendChild(tingleScriptElement);
-
-        tingleScriptElement.onload = () => {
-          const modal = new tingle.modal({
-            footer: true,
-          });
-
-          modal.setContent("<p>このサイトでは、センサー値を扱います。</p>");
-          modal.addFooterBtn(
-            "Cancel",
-            "tingle-btn tingle-btn--default tingle-btn--pull-right",
-            () => modal.close(),
-          );
-          modal.addFooterBtn(
-            "OK",
-            "tingle-btn tingle-btn--primary tingle-btn--pull-right",
-            () => {
-              modal.close();
-            },
-          );
-
-          document
-            .querySelector(
-              ".tingle-btn.tingle-btn--primary.tingle-btn--pull-right",
-            )
-            ?.addEventListener("click", async () => {
-              // eslint-disable-next-line @typescript-eslint/no-explicit-any
-              const isDeviceOrientationEvent = await (
-                DeviceOrientationEvent as any
-              ).requestPermission();
-              // eslint-disable-next-line @typescript-eslint/no-explicit-any
-              const isDeviceMotionEvent = await (
-                DeviceMotionEvent as any
-              ).requestPermission();
-
-              // 許可したあとはまた許可が必要になるまでポップアップが出ないようにする
-              if (isDeviceOrientationEvent && isDeviceMotionEvent) {
-                sessionStorage.setItem("isPermission", "true");
-              }
-            });
-
-          modal.open();
-        };
-      };
-    }
-  }, []);
 
   useEffect(() => {
     const orientationHandler = (orientation: DeviceOrientationEvent) => {
@@ -192,6 +129,8 @@ export const MobileSensorData: React.VFC = () => {
           <List values={rotationRateLists} />
         </Section>
       </CardArticle>
+
+      <IOSPermissionsCheckModal />
     </>
   );
 };
